remove watchers, update on select events
