+++ src/composables/useDiscogs.ts
@@ -1,7 +1,7 @@
 import { readonly, ref } from 'vue'
 import { type ResponseData } from '@/types/ResponseData'
-import { genres, type Genre } from '@/types/Genre'
-import { decades, type Decade } from '@/types/Decade'
+import { type Genre } from '@/types/Genre'
+import { type Decade } from '@/types/Decade'
 import type { AlbumData } from '@/types/AlbumData'
 
 import getYear from './getYear'
@@ -12,16 +12,6 @@ const selected_album = ref<AlbumData | null>(null)
 const error = ref<Error | null>(null)
 
 export default function useDiscogs() {
-  const getRandomGenre = () => {
-    const genre_index = Math.floor(Math.random() * genres.length)
-    return genres[genre_index]
-  }
-
-  const getRandomDecade = () => {
-    const decade_index = Math.floor(Math.random() * decades.length)
-    return decades[decade_index]
-  }
-
   const get_fetch_url = (genre: Genre, year: number) => {
     const url = new URL('https://api.discogs.com/database/search')
     url.searchParams.set('type', 'master')
@@ -50,42 +40,26 @@ export default function useDiscogs() {
     return album_data
   }
 
-  const fetchAlbums = async (genre: Genre, year: number) => {
-    const controller = new AbortController()
-
-    const fetch_url = get_fetch_url(genre, year)
-    const response = await fetch(fetch_url, {
-      headers: {
-        Authorization: 'Discogs key=SJdGocnZCcCymeDLaWgC, secret=hVVwFIdXbknmrXffmLCdvFUKiQSHaQvf'
-      },
-      signal: controller.signal
-    })
-    if (!response.ok) {
-      throw Error('Error retrieving album data')
-    }
-
-    const queried_data = await response.json()
-
-    return queried_data.results as ResponseData[]
-  }
-
-  /**
-   * Loads a new album into `selected_album`. If any option is not specified
-   * it is chosen randomly from all its options.
-   *
-   * @param options fetch options
-   */
-  const refreshAlbum = async (options: { genre?: Genre; decade?: Decade } = {}) => {
-    let { genre, decade } = options
-
-    if (!genre) genre = getRandomGenre()
-    if (!decade) decade = getRandomDecade()
-
-    // Get year from decade
-    const year = getYear(decade)
-
+  const fetchAlbum = async (genre: Genre, decade: Decade) => {
     try {
-      const responseData = await fetchAlbums(genre, year)
+      // Get year from decade
+      const year = getYear(decade)
+
+      const controller = new AbortController()
+
+      const fetch_url = get_fetch_url(genre, year)
+      const response = await fetch(fetch_url, {
+        headers: {
+          Authorization: 'Discogs key=SJdGocnZCcCymeDLaWgC, secret=hVVwFIdXbknmrXffmLCdvFUKiQSHaQvf'
+        },
+        signal: controller.signal
+      })
+      if (!response.ok) {
+        throw Error('Error retrieving album data')
+      }
+
+      const queried_data = await response.json()
+      const responseData = queried_data.results as ResponseData[]
       selected_album.value = getAlbumDataFromResponse(responseData, genre, year)
     } catch (err) {
       if (err instanceof Error) {
@@ -99,8 +73,6 @@ export default function useDiscogs() {
     selected_album: readonly(selected_album),
     error: readonly(error),
 
-    getRandomDecade,
-    getRandomGenre,
-    refreshAlbum
+    fetchAlbum
   }
 }
